backend-django/requirements.txt
Django
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
PyMySQL==1.1.1
drf-yasg
django-prometheus